Fix: change age to int.

+++ src/main/java/com/aerospike/demo/simplespringbootaerospikedemo/objects/User.java
@@ -13,5 +13,5 @@ public class User {
     private int id;
     private String name;
     private String email;
-    private short age;
+    private int age;
 }
